Deployment/requirements.txt
numpy==1.19.5
tensorflow==2.6.2
keras==2.6.0
opencv==4.1.2
pillow==7.0.0
streamlit==1.8.1
